scripts/requirements.txt
autopep8==1.5.6
enlighten==1.10.2
iso3166==2.0.2
Mako==1.1.6
mwclient==0.10.1
openpyxl==3.0.9
prettytable==2.4.0
prompt-toolkit==3.0.24
python-dotenv==0.15.0
requests==2.25.1
